+++ babel.config.json
@@ -1,3 +1,12 @@
 {
-  "presets": ["@babel/preset-env"]
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "node": "9.2.1"
+        }
+      }
+    ]
+  ]
 }
