+++ public/css/main.css
@@ -529,7 +529,7 @@ button {
 }
 .film-details__title {
   font-size: 50px;
-  margin: 0;
+  margin: 0 0 12px;
 }
 .film-details__title-original {
   font-size: 21px;
@@ -707,7 +707,7 @@ button {
 }
 .film-details__comment-text {
   margin: 0;
-  margin-bottom: 5px;
+  margin-bottom: 10px;
   font-size: 20px;
   font-weight: 600;
 }
