+++ src/mock/card-data.js
@@ -1,0 +1,136 @@
+const ACTORS = [
+  'Takeshi Kitano',
+  'Morgan Freeman ',
+  'Brad Pitt',
+  'Al Pacino',
+  'Robert De Niro',
+  'Tom Hanks',
+  'Edward Norton',
+  'Christian Bale',
+  'Ralph Fiennes',
+  'Cillian Murphy',
+  'Michael Caine',
+  'Gary Oldman',
+  'Harrison Ford',
+  'Leonardo DiCaprio',
+  'Matt Damon'
+];
+
+const COUNTRIES = [
+  'USA',
+  'Germany',
+  'China',
+  'France',
+  'Japan',
+  'Spain',
+  'Russia',
+  'Italy',
+  'Finland'
+];
+
+const DESCRIPTION_PARTS = [
+  'oscar-winning film',
+  'a war drama about two young people',
+  'true masterpiece where love and death are closer to heroes than their family',
+  'from the creators of timeless classic "Nu Pogodi!" and "Alice in Wonderland"',
+  'a film about a journey that heroes are about to make in finding themselves',
+  'with the best fight scenes since Bruce Lee',
+];
+
+const DIRECTORS = [
+  'Alejandro Gonsales Inarritu',
+  'Brad Bird',
+  'Tom Ford',
+  'Akira Kurosawa',
+  'James Cameron',
+  'Clint Eastwood',
+  'Quentin Tarantino',
+  'Chrostopher Nolan'
+];
+
+const GENRES = [
+  'Adventure',
+  'Drama',
+  'Comedy',
+  'Action',
+  'Horror',
+  'Animation',
+  'Family',
+  'Thriller',
+  'Sci-Fi'
+];
+
+const POSTERS = [
+  'made-for-each-other.png',
+  'popeye-meets-sinbad.png',
+  'sagebrush-trail.jpg',
+  'santa-claus-conquers-the-martians.jpg',
+  'the-dance-of-life.jpg',
+  'the-great-flamarion.jpg',
+  'the-man-with-the-golden-arm.jpg',
+];
+
+const TITLE_BEGINNINGS = [
+  'A Lion',
+  'A Man',
+  'A Shark',
+  'A Tale Of A Little Bird',
+  'Country',
+  'Family',
+  'Guest',
+  'Happiness',
+  'Laziness',
+  'Pioneers',
+  'Raiders',
+];
+
+const TITLE_MIDDLES = [
+  'In',
+  'Of',
+  'On',
+  'Who Bought',
+  'Who Saw',
+  'Who Sold',
+  'Who Stole',
+  'With',
+  'Within',
+  'Without',
+];
+
+const TITLE_ENDINGS = [
+  'Him',
+  'Himself',
+  'The Carpet',
+  'The Darkness',
+  'The Floor',
+  'The Room',
+  'The Storm',
+  'The Void',
+  'The Wall',
+  'Themselves',
+  'Us',
+];
+
+const WRITERS = [
+  'Quentin Tarantino',
+  'Hayao Miazaki',
+  'Takeshi Kitano',
+  'Martin Scorsese',
+  'Robert Rodrigues',
+  'Robert Zemeckis',
+  'Stephen Spielberg',
+  'Brad Bird'
+];
+
+export {
+  ACTORS,
+  COUNTRIES,
+  DESCRIPTION_PARTS,
+  DIRECTORS,
+  GENRES,
+  POSTERS,
+  TITLE_BEGINNINGS,
+  TITLE_MIDDLES,
+  TITLE_ENDINGS,
+  WRITERS,
+};
