+++ src/mock/card.js
@@ -1,163 +1,39 @@
 import {
-  getIdGenerator,
+  idGenerator,
   getRandom,
   getRandomDate,
   getRandomElementFrom,
   getRandomInt,
-  getTextGenerator,
+  getTextGenerator, getUniqueRandomFromRange, getUniqueRandomFromArrayGenerator,
 } from '../utils';
+import CommentsModel from '../model/comments-model';
+import {
+  ACTORS,
+  COUNTRIES,
+  DESCRIPTION_PARTS,
+  DIRECTORS,
+  GENRES,
+  POSTERS,
+  TITLE_BEGINNINGS,
+  TITLE_MIDDLES,
+  TITLE_ENDINGS,
+  WRITERS
+} from './card-data.js';
 
-const ACTORS = [
-  'Takeshi Kitano',
-  'Morgan Freeman ',
-  'Brad Pitt',
-  'Al Pacino',
-  'Robert De Niro',
-  'Tom Hanks',
-  'Edward Norton',
-  'Christian Bale',
-  'Ralph Fiennes',
-  'Cillian Murphy',
-  'Michael Caine',
-  'Gary Oldman',
-  'Harrison Ford',
-  'Leonardo DiCaprio',
-  'Matt Damon'
-];
+const id = idGenerator();
 
-const COUNTRIES = [
-  'USA',
-  'Germany',
-  'China',
-  'France',
-  'Japan',
-  'Spain',
-  'Russia',
-  'Italy',
-  'Finland'
-];
-
-const DESCRIPTION_PARTS = [
-  'oscar-winning film',
-  'a war drama about two young people',
-  'true masterpiece where love and death are closer to heroes than their family',
-  'from the creators of timeless classic "Nu Pogodi!" and "Alice in Wonderland"',
-  'a film about a journey that heroes are about to make in finding themselves',
-  'with the best fight scenes since Bruce Lee',
-];
-
-const DIRECTORS = [
-  'Alejandro Gonsales Inarritu',
-  'Brad Bird',
-  'Tom Ford',
-  'Akira Kurosawa',
-  'James Cameron',
-  'Clint Eastwood',
-  'Quentin Tarantino',
-  'Chrostopher Nolan'
-];
-
-const GENRES = [
-  'Adventure',
-  'Drama',
-  'Comedy',
-  'Action',
-  'Horror',
-  'Animation',
-  'Family',
-  'Thriller',
-  'Sci-Fi'
-];
-
-const POSTERS = [
-  'made-for-each-other.png',
-  'popeye-meets-sinbad.png',
-  'sagebrush-trail.jpg',
-  'santa-claus-conquers-the-martians.jpg',
-  'the-dance-of-life.jpg',
-  'the-great-flamarion.jpg',
-  'the-man-with-the-golden-arm.jpg',
-];
-
-const TITLE_BEGINNINGS = [
-  'A Lion',
-  'A Man',
-  'A Shark',
-  'A Tale Of A Little Bird',
-  'Country',
-  'Family',
-  'Guest',
-  'Happiness',
-  'Laziness',
-  'Pioneers',
-  'Raiders',
-];
-
-const TITLE_MIDDLES = [
-  'In',
-  'Of',
-  'On',
-  'Who Bought',
-  'Who Saw',
-  'Who Sold',
-  'Who Stole',
-  'With',
-  'Within',
-  'Without',
-];
-
-const TITLE_ENDINGS = [
-  'Him',
-  'Himself',
-  'The Carpet',
-  'The Darkness',
-  'The Floor',
-  'The Room',
-  'The Storm',
-  'The Void',
-  'The Wall',
-  'Themselves',
-  'Us',
-];
-
-const WRITERS = [
-  'Quentin Tarantino',
-  'Hayao Miazaki',
-  'Takeshi Kitano',
-  'Martin Scorsese',
-  'Robert Rodrigues',
-  'Robert Zemeckis',
-  'Stephen Spielberg',
-  'Brad Bird'
-];
-
-const COMMENTS_IDS_RANGE = 50;
-
-const generateId = getIdGenerator();
-
-const getCommentsIdGenerator = (range) => {
-  const generatedIds = new Set();
-  return () => {
-    let id;
-    do {
-      id = getRandomInt(0, range - 1).toString();
-      if (generatedIds.size >= range) {
-        throw new Error('Comment id range limit reached');
-      }
-    } while (generatedIds.has(id));
-    generatedIds.add(id);
-    return id;
-  };
-};
-
-const generateCommentsId = getCommentsIdGenerator(COMMENTS_IDS_RANGE);
+const commentId = getUniqueRandomFromRange(CommentsModel.getCommentsCount());
 
 const generateCard = () => {
   const generateDescription = getTextGenerator(DESCRIPTION_PARTS);
+  const getRandomWriter = getUniqueRandomFromArrayGenerator(WRITERS);
+  const getRandomActor = getUniqueRandomFromArrayGenerator(ACTORS);
+  const getRandomGenre = getUniqueRandomFromArrayGenerator(GENRES);
+
   return {
-    id: generateId().toString(),
+    id: id.next().value.toString(),
     comments: Array
-      .from({length: getRandomInt(1, 5)}, generateCommentsId),
+      .from({length: getRandomInt(1, 5)}, () => commentId.next().value.toString()),
     filmInfo: {
       title: `${getRandomElementFrom(TITLE_BEGINNINGS)} ${getRandomElementFrom(TITLE_MIDDLES)} ${getRandomElementFrom(TITLE_ENDINGS)}`,
       alternativeTitle: `${getRandomElementFrom(TITLE_BEGINNINGS)} ${getRandomElementFrom(TITLE_MIDDLES)} ${getRandomElementFrom(TITLE_ENDINGS)}`,
@@ -165,14 +41,14 @@ const generateCard = () => {
       poster: `images/posters/${getRandomElementFrom(POSTERS)}`,
       ageRating: getRandomInt(0, 18),
       director: getRandomElementFrom(DIRECTORS),
-      writers: Array.from({length: getRandomInt(1, 3)}, () => getRandomElementFrom(WRITERS)),
-      actors: Array.from({length: getRandomInt(3, 6)}, () => getRandomElementFrom(ACTORS)),
+      writers: Array.from({length: getRandomInt(1, 3)}, getRandomWriter),
+      actors: Array.from({length: getRandomInt(3, 6)}, getRandomActor),
       release: {
         date: getRandomDate(new Date(1900, 0, 1), new Date()).toISOString(),
         releaseCountry: getRandomElementFrom(COUNTRIES),
       },
       runtime: getRandomInt(30, 180),
-      genre: Array.from({length: getRandomInt(1, 3)}, () => getRandomElementFrom(GENRES)),
+      genre: Array.from({length: getRandomInt(1, 3)}, getRandomGenre),
       description: generateDescription(3, 5),
     },
     userDetails: {
