+++ src/presenter/cards-container-presenter.js
@@ -1,9 +1,9 @@
 import CardsExtraListView from '../view/cards-extra-list-view';
 import CardsListPresenter from './cards-list-presenter';
 import CardsListView from '../view/cards-list-view';
+import DetailsPresenter from './details-presenter';
 import MoreButtonView from '../view/more-button-view';
 import {render} from '../render';
-import DetailsPresenter from './details-presenter';
 
 export default class CardsContainerPresenter {
 
@@ -11,7 +11,7 @@ export default class CardsContainerPresenter {
     this.cardsModel = cardsModel;
     this.commentsModel = commentsModel;
     this.cards = [...this.cardsModel.getCards()];
-    this.comments = [...this.commentsModel.getComments()];
+
 
     const cardsListView = new CardsListView();
     const topListView = new CardsExtraListView();
@@ -26,7 +26,8 @@ export default class CardsContainerPresenter {
     new CardsListPresenter().init(topListView, this.cards.slice(0, 2));
     new CardsListPresenter().init(popularListView, this.cards.slice(0, 2));
 
-    const detailsPresenter = new DetailsPresenter(this.cards, this.comments);
+
+    const detailsPresenter = new DetailsPresenter(this.cards);
     detailsPresenter.init(container);
   }
 }
